Update iOSXenkoView.cs
When implementing full page or interstitial ads such as google admob a black screen appears on any iOS device after the ad is loaded. This is a known problem with OpenTK-1.1 and the solution is to override WillMoveToWindow in iPhoneOSGameView so as to not suspend the window which causes the problem.