Updated Package Support Checks
Checking for Android, iOS, and UWP packages should now be up-to-date for VS2017 and still work with VS2015.

I tried to make it somewhat expandable via Version-to-path dictionaries. They won't need to be updated unless a package path changes.

I... Couldn't find a better iOS package to check for. That was the only one to mention iOS, unless that's rolled into Xamarin.VisualStudio.Apple.Sdk? Any suggestions for better packages would be much appreciated!