[Math] Add Single2/3/4, Double2/3/4, make Vectors etc use real_t
Xenko is an engine designed for the future. You know what is a great feature to have for future games? Double precision! This commit does two things. One, it adds Single2/3/4 and Double2/3/4 classes, for when you explicitly need to work with either Single or Double precision. These have conversions between each other and Vectors. Two, this commit changes "float" in several math types to "real_t". The easiest way to support doubles, at least for transitioning, is to support Single and Double precision simultaneously. This can be done in C# with "using" and preprocessor directives. This commit abstracts away the "float" variables in a few classes to a new variable called "real_t", which for now can only represent "float".
g