Add IDebugPrimitive instead of using MeshDraw at CreateDebugPrimitive
The purpose of this is preparation to implement debug primitive for HeightfieldColliderShape.
PhysicsShapesRenderingServices gets multiple MeshDraw from IDebugPrimitive.
IDebugPrimitive is also used as args to update debug primitive at runtime.

Heightfield's debug primitive will easily get over the limit of indices in Level_9_x of GraphicsProfile.
Multiple meshes are needed to create a model for that.
And that shape will be updated at runtime.