Fix to stop the visual studio installer exiting with code -1 during prerequisites install
The prerequisites installer somehow keeps the msiexec service in an installing state even though the installer is already finished and closed. This causes the visual studio installer to refuse to install because it thinks there is still an other installer running.

By installing the visual studio prerequisite before installing the other prequisites this problem is avoided. This is a hack because the windows installer state will still be wrong after this, but this will at least allow people to install stride properly without encountering an error that they can't easily solve. The windows installer being in a wrong state will most likely not cause any other problems for the user, because the next windows installer will detect the wrong state and fix it, or in the case of the visual studio installer it will fix it after hitting retry or restarting the visual studio installer.

The proper solution for not ending up in this state would be in the advanced installer script/version. However, I could not find any online solutions or other people with the same problem.