Fixes a dead lock in the effect system
The `DynamicEffectInstance` does a blocking wait on the result of the effect system. In there a continuation is created in case the compilation task is not yet finished. In case that continuation got scheduled on the calling thread we ended up in a dead lock.

See also https://github.com/TheFuseLab/VL.Fuse/issues/51