Updated the readme to call out msbuild instead of dotnet.
The dotnet command can't compile C++, which causes build failures. msbuild from the build tool (or vs) just fine on its own.

The previous step adds msbuild to the path, so this should be fine. To my knowedge dotnet is doing this same thing, just it doesn't have the libraries accessable to the msbuild install at the added-path-location.