Fix for issue #1185 - Stride.Launcher - Could not find a part of the path
The launcher executable was moved to a new tools directory as part of the fix for #1139.  Older packages did not contain a tools directory so added code to check for the existence of a target directory before delving into it.

Also tightened up the code in the affected area and removed an unused using statement.