Next round of addressing Warnings
This change addresses most or all of the follwing issues:
CS0414 - The field '<foo>' is assigned but its value is never used
CS0612 - <member> is obsolete.   (see Note below)
CS0649 - Field '<foo>' is never assigned to, and will always have its default value
CS0659 - '<foo>' overrides Object.Equals(object o) but does not override Object.GetHashCode()
CS0675 - Bitwise-or operator used on a sign-extended operand; consider casting to a smaller unsigned type first
CS1998 - This async method lacks 'await' operators and will run synchronously.
CS4014 - Because this call is not awaited, execution of the current method continues before the call is completed.
CS8073 - The result of the expression is always 'false' since a value of type 'Guid' is never equal to 'null' of type 'Guid?'

Note that beacuse NodeViewModel.MemberInfo was always null, this caused 3 dependent providers to always have their MatchNode method return false.   I left the three classes in this iteration beacuse they are still referenced in xaml and I'm unclear of the effect of removing the xaml at this time.