Move or remove problematic code
Related to duplicated code causing compiler issues, this change either moves or removes duplicate references to shared code within the same end application.

Some of the shared code turns out to be obsolete related to live debugging in Studio.  In this case, this change also removes those classes as well.