[Physics] Adding margins to convex hull colliders (Issue1577) (#2257)
* Added margin fields to ConvexHullCollider (#1577)

Added margin field to parent class ColliderShape, therefore all collider shapes can have access.

* Added default for margin to be 0.04f

Context: Margins in bullet by default should be around 0.04 for dynamic and 0.0 for static objects. Convex hulls are likely used as replacement for the lack of support of non-static meshes, so best set it to the default value for dynamics, so 0.04 by default.