Clarify that MSBuild is required in most cases
----
Some command line examples mention using `dotnet build`. However, unless when building Core libraries, engine and editor requires the use of msbuild because of a few C++/CLI projects.
This should be made clear in the documentation.

Correct the command line examples throughout the documentation to use msbuild when required.